api_gateway/requirements.txt
# api_gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.12
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.24.0
redis==4.5.5
python-dotenv==1.0.0
